+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Added
 * In learn.py, DFData does train/test split.
 * NN Trainer does validation on holdout test data set.
 * Can now transfer Study data to an archive dict.
+* Method for deleting orphaned GridFS file data.
 
 Fixed
 -----